pip-requirements.txt
psycopg2==2.7.5
celery[redis]>=4.1.1,<4.2.0
apache-airflow[postgres,celery,crypto]==1.10.0